move lead component and accompanying files to separate folder
