+++ client/src/components/Lead/Lead.test.js
@@ -2,7 +2,7 @@ import React from 'react'
 import {configure, shallow} from 'enzyme';
 import Adapter from 'enzyme-adapter-react-16'
 import Lead from './Lead'
-import profile from '../img/profile.svg'
+import profile from '../../img/profile.svg'
 
 
 configure({adapter: new Adapter});
